+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (Unreleased)
 
 - Add new cop `RSpec/LeakyLocalVariable`. ([@lovro-bikic])
+- Add new cop `RSpec/RedundantPending`. ([@ydah])
 - Bump RuboCop requirement to +1.81. ([@ydah])
 - Fix a false positive for `RSpec/LetSetup` when `let!` used in outer scope. ([@ydah])